requirements.txt
# Core
python-dotenv>=1.0.0

# LangChain + LangGraph
langchain>=0.2.10
langgraph>=0.2.30
langchain-core>=0.2.30
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2

# Google Gemini
langchain-google-genai>=1.0.8

# Vector store
chromadb>=0.5.5

# PDF + Images
pymupdf>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
pymupdf 
pillow 
pytesseract

flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1